Bug 1493439 - Remove the canI check for secrets from bind-result
The `canI` check only works if you've visited the project. This is not
always true when provisioning from the catalog, so it means the secret
is sometimes not displayed as a link when it should be.

To provision a service or create a binding with parameters, it's
required that the user create secrets anyway, so the check should never
fail is permissions are set up properly.

If the user doesn't have authority, the worst that happens is that they
see a permission denied error following the link. In practice, it
shouldn't happen.

https://bugzilla.redhat.com/show_bug.cgi?id=1493439